requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.3
numpy>=1.21.2
scikit-learn>=0.24.2
firebase-admin>=5.0.3
python-dotenv>=0.19.1
pydantic>=1.8.2
joblib>=1.0.1
